requirements.txt
python-openhab==2.14.0
inky==0.0.5
buttonshim==0.0.2